# perf(dashboards): improve API performance for dashboards