--- a/setup.py
+++ b/setup.py
@@ -76,7 +76,7 @@ setup(
         "croniter>=0.3.28",
         "cryptography>=2.4.2",
         "flask>=1.1.0, <2.0.0",
-        "flask-appbuilder>=2.3.2, <2.4.0",
+        "flask-appbuilder>=2.3.4, <2.4.0",
         "flask-caching",
         "flask-compress",
         "flask-talisman",