### CATEGORY

- [ ] Bug Fix
- [X] Enhancement (new features, refinement)
- [ ] Refactor
- [ ] Add tests
- [ ] Build / Development Environment
- [ ] Documentation

### SUMMARY
Uses new FAB 2.3.4 that improves performance on table joins. Again like stated on #9619, when using properties has virtual columns, we should to make sure that the method references columns already referenced on `list_columns` or SQLAlchemy will issue a new query

Note: We should merge #9703 first

Before:
[stats_logger] (timing) DashboardRestApi.get_list.time 110ms - 150ms

After:
[stats_logger] (timing) DashboardRestApi.get_list.time 15ms - 30ms

### ADDITIONAL INFORMATION
- [ ] Has associated issue:
- [ ] Changes UI
- [ ] Requires DB Migration.
- [ ] Confirm DB Migration upgrade and downgrade tested.
- [ ] Introduces new feature or API
- [ ] Removes existing feature or API

### REVIEWERS
@willbarrett @nytai